+++ src/pyoak/match/__init__.py
@@ -1,0 +1,8 @@
+from .pattern import from_pattern, validate_pattern
+from .xpath import ASTXpath
+
+__all__ = [
+    "ASTXpath",
+    "from_pattern",
+    "validate_pattern",
+]