+++ tests/pyoak/match/test_pattern.py
@@ -4,9 +4,9 @@ from dataclasses import dataclass, field
 from typing import Any
 
 import pytest
+from pyoak.match import from_pattern, validate_pattern
 from pyoak.match.error import ASTXpathOrPatternDefinitionError
 from pyoak.match.matcher import NodeMatcher
-from pyoak.match.pattern import from_pattern, validate_pattern
 from pyoak.node import ASTNode
 from pyoak.origin import NO_ORIGIN, Origin